Moved one test back to CompactionIT
